refs TTTS-477 [POS連携]timeoutエラーにも関わらず完了フォルダにファイルが移動する

+++ merge_function/index.js
@@ -48,7 +48,7 @@ if (process.env.NODE_ENV !== 'production') {
 module.exports = (context, myBlob) => __awaiter(this, void 0, void 0, function* () {
     //export async function run (context, myBlob) {
     context.funcId = context.executionContext.invocationId;
-    context.log(`ファイル: ${context.funcId}`);
+    context.log(`ProcessId: ${context.funcId}`);
     try {
         mongoose.connect(process.env.MONGOLAB_URI, configs.mongoose);
         const rows = yield readCsv(context);
@@ -154,8 +154,34 @@ function moveListFileWorking(context) {
     return __awaiter(this, void 0, void 0, function* () {
         const oriBlob = 'working/' + context.bindingData.name;
         const targetBlob = 'complete/' + context.bindingData.name;
-        yield storage.createBlobService().startCopyBlob(context.bindingData.uri + '?sasString', process.env.AZURE_BLOB_STORAGE, targetBlob, (error, result, res) => __awaiter(this, void 0, void 0, function* () {
-            yield storage.createBlobService().deleteBlobIfExists(process.env.AZURE_BLOB_STORAGE, oriBlob, (error, result, res) => __awaiter(this, void 0, void 0, function* () { }));
-        }));
+        const tableName = 'AzureWebJobsHostLogs' + moment(moment().toISOString()).format('YYYYMM');
+        const tableService = storage.createTableService();
+        const checkTableExists = () => __awaiter(this, void 0, void 0, function* () {
+            return new Promise((resolve, rejects) => {
+                tableService.createTableIfNotExists(tableName, function (error, result, response) {
+                    resolve(result);
+                });
+            });
+        });
+        const checkTbl = yield checkTableExists();
+        if (checkTbl.isSuccessful === true) {
+            const query = new storage.TableQuery().where('RowKey == ?', context.funcId);
+            const getProcessInformation = () => __awaiter(this, void 0, void 0, function* () {
+                return new Promise((resolve, reject) => {
+                    tableService.queryEntities(tableName, query, null, (err, data) => {
+                        if (err)
+                            reject(err);
+                        else
+                            resolve(data.entries.length > 0 ? data.entries[0] : undefined);
+                    });
+                });
+            });
+            const processInformation = yield getProcessInformation();
+            if (processInformation.ErrorDetails === undefined) {
+                yield storage.createBlobService().startCopyBlob(context.bindingData.uri + '?sasString', process.env.AZURE_BLOB_STORAGE, targetBlob, (error, result, res) => __awaiter(this, void 0, void 0, function* () {
+                    yield storage.createBlobService().deleteBlobIfExists(process.env.AZURE_BLOB_STORAGE, oriBlob, (error, result, res) => __awaiter(this, void 0, void 0, function* () { }));
+                }));
+            }
+        }
     });
 }
